feat: Add `infer` inference to `createHands`
If I'm not mistaken, this single change to `createHands` should do the
exact same thing as the main `infer` function (with the exception of
editing the suggestions list, of course).

The suggestions list should still be updated to show info to the user in
the game log, but it doesn't need to be part of the inference
process--it can just be done afterwards.

Using this change should also make `probabilities` significantly more
performant.