feat: Strip inference logic from `infer` + rename
The next commit will simply change `createHands`'s name to `infer`, but
I thought it'd be best to split the change so that there isn't a huge
amount of additions/deletions as well as making one function take the
name of a previous one.

NOTE: This is technically a small regression because of the removal of
the code for Clue variations in which multiple players can reveal at
once. I'll decide whether it's worth it to add this back with the new
system (eg. with regards to performance hits).