feat: Add inference using multiple maybeGroups
If there are multiple maybeGroups that are disjoint from `hand.has`,
then the remaining cards in the player's hand must be elements of the
maximum disjoint set of the eligible maybeGroups. This check is not
necessary when the player only has one unknown card, in which case the
original code is used.