feat: Switch to grid layout
While still not beautiful, this does make the site substantially less
ugly. It also now works properly on mobile! The only issue is that in
its current state, the user has to manually zoom out in order to fit the
website onto their screen. I'll see if there's any good way to do this
automatically.