fix(probabilities): Extra condition before marking occurence
Before this change, the function would stop exploring a possibility as
soon as a suspect, weapon, and room were found. This actually resulted in
skewed odds, since it didn't fully consider all of the different hand
arrangements that could contribute to a given suspect/weapon/room.

With this change, odds should become more accurate.