fix(rust): Switch from mini-alloc to wee_alloc
I somehow missed the part of mini-alloc's memory that mentions that it
deliberately leaks memory in order to improve performance.
Unfortunately, when calculating odds, a memory leak doesn't actually
take all that long to happen. This allocator doesn't leak memory afaik,
and it is similarly performant (though of course it is a little bit
slower).