perf: Remove unneeded reassignments
Svelte's Proxies mean that doing things like `arr = [...arr, new]` or
`arr.push(new); arr=arr` are no longer required, so this commit removes
such lines. The only instance that was kept is for the `$sets` store,
because it is a `Map` and Svelte does not properly proxy its methods.