feat: Remove maybe groups with found cards
Closes #8

This should help reduce clutter on the clue sheet. This also has the
side effect of speeding up the inference function a decent bit (easiest
to notice when running `probabilities`) since: 1) it takes less effort
to filter eligible groups for a couple inference methods, and 2) the
check to see if it's worth running the expensive computation will now be
true less often.