refactor: Stop using maps for maybe groups
I realized that there's actually no point in using a map to store maybe
groups, so it's just been changed to a `Vec`. Surprisingly, this didn't
improve performance. But it _did_ reduce the binary size by 10 KB!

Also, the test files are built expecting maybe groups to be a `Map`, so
for the time being I've just updated the test runner to convert this
into a plain list instead. Tests all pass without any other changes.