feat: Add option to disable inference
This lets the website essentially just act as a public log of
suggestions. It could be useful in a game where people want to play
normally (i.e. keeping their own paper clue sheets), and want a way to
audit suggestions after the game is over to make sure that everybody
showed cards when they were supposed to.

It can also be interesting to see how much data the site was able to
figure out just from what people suggested and the responses!