perf(inference): Use more granular comparisons to detect changes
Instead of cloning the list of hands every iteration and then seeing if
it changed, the code now only compares properties affected by a given
inference method to detect changes.

For example, if an inference technique might update a player's `missing`
set, then only the original `missing` set is copied and compared to the
state after the technique runs.

Since sets here are just `u64`s, copies are essentially free, unlike
the original method which cloned the entire `Vec` of player hands.